Test/ez 91 stat df (#41)
* refactor: extract test_report

* refactor: sbr fixture

* refactor: remove propertymock

* test: add test_stat_df

* test: test_pct_maximum_drawdown

* test: test_maximum_drawdown

* fix: ignore warning

* fix: count_true_consecutive
