+++ ezyquant/utils.py
@@ -132,6 +132,15 @@ def _date_range(
     return out
 
 
+def count_true_consecutive(s: pd.Series) -> int:
+    """Count the number of consecutive True values in a series."""
+    out = (s * (s.groupby((s != s.shift()).cumsum()).cumcount() + 1)).max()
+    if pd.isna(out):
+        return 0
+    else:
+        return out
+
+
 """
 Cache
 """
