+++ tests/test_utils.py
@@ -286,6 +286,32 @@ class TestIsRebalanceMonthly:
         assert_series_equal(result, expected_s)
 
 
+@pytest.mark.parametrize(
+    ("series", "expect_result"),
+    [
+        (pd.Series([]), 0),
+        (pd.Series([False]), 0),
+        (pd.Series([True]), 1),
+        (pd.Series([False, False]), 0),
+        (pd.Series([False, True]), 1),
+        (pd.Series([True, False]), 1),
+        (pd.Series([True, True]), 2),
+        (pd.Series([False, False, False]), 0),
+        (pd.Series([False, False, True]), 1),
+        (pd.Series([False, True, False]), 1),
+        (pd.Series([False, True, True]), 2),
+        (pd.Series([True, False, False]), 1),
+        (pd.Series([True, False, True]), 1),
+        (pd.Series([True, True, False]), 2),
+        (pd.Series([True, True, True]), 3),
+    ],
+)
+def test_count_true_consecutive(series: pd.Series, expect_result):
+    result = utils.count_true_consecutive(series)
+    print(result)
+    assert result == expect_result
+
+
 class TestWrapCacheClass:
     def _test(
         self,
